Reorganize signup
